Let the dust settle on the connection refactor
Moves connection handling into dedicated modules and simplifies the
Connection module by removing redundant code and reorganizing existing
logic. This hopefully makes the connection management easier to
understand.
