+++ lib/familia/horreum/core/serialization.rb
@@ -1,5 +1,7 @@
 # lib/familia/horreum/serialization.rb
-#
+
+require_relative '../../multi_result'
+
 module Familia
   # Familia::Horreum
   #
